ProyectoCasiTerminado

+++ Api/Controllers/AlumnoCarreraController.cs
@@ -51,7 +51,7 @@ namespace Api.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IResult> Eliminar(long id)
+        public async Task<IResult> Eliminar(int id)
         {
             await _alumnoCarreraServicio.Eliminar(id);
