+++ Api/PersistenceModels/CuotaModel.cs
@@ -4,8 +4,8 @@
     {
         public int Numero { get; set; }
 
-        public long AlumnoId { get; set; }
+        public int AlumnoId { get; set; }
 
-        public long PrecioCuotaId { get; set; }
+        public int PrecioCuotaId { get; set; }
     }
 }
