+++ Dominio/Entidades/Pago.cs
@@ -13,7 +13,7 @@ namespace Dominio.Entidades
 
         public decimal PorcPago { get; set; }
 
-        public long CuotaId { get; set; }
+        public int CuotaId { get; set; }
 
         public DateTime FechaPago { get; set; }
