+++ Dominio/Metadata/ICuota.cs
@@ -27,11 +27,11 @@ namespace Dominio.Metadata
 
         [Required(ErrorMessage = "El campo {0} es Obligatorio.")]
         [Display(Name = "Alumno")]
-        long AlumnoId { get; set; }
+        int AlumnoId { get; set; }
 
         [Required(ErrorMessage = "El campo {0} es Obligatorio.")]
         [Display(Name = "PrecioCuota")]
-        long PrecioCuotaId { get; set; }
+        int PrecioCuotaId { get; set; }
 
     }
 }
