+++ Dominio/Metadata/IProceso.cs
@@ -26,7 +26,7 @@ namespace Dominio.Metadata
 
         [Required(ErrorMessage = "El campo {0} es Obligatorio.")]
         [Display(Name = "Usuario")]
-        long UsuarioId { get; set; }
+        int UsuarioId { get; set; }
 
     }
 }
