+++ IServicios/Usuario/UsuarioDTO/UsuarioDTO.cs
@@ -9,7 +9,7 @@ namespace IServicios.Usuario.UsuarioDTO
         public string Nombre { get; set; }
         public string Password { get; set; }
         public Rol Rol { get; set; }
-        public long PersonaId { get; set; }
+        public int PersonaId { get; set; }
         public PersonaDTO Persona { get; set; }
     }
 }
